# new-Health/src/api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Health Microservice deals with ECG, BP, Sugar, Stress and other Parameters",
        "version": "0.0.1",
        "title": "Health Microservice"
    },
    "host": "localhost:1000",
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json",
        "multipart/form-data"
    ],
    "produces": [
        "application/json"
    ],
    "securityDefinitions": {
        "basicAuth": {
            "type": "basic"
        }
    },
    "paths": {
...
        "/getstressofdeviceid/{deviceId}": {
            "x-swagger-router-controller": "stressController",
            "get": {
                "tags": [
                    "Stress"
                ],
                "operationId": "getStressByDeviceId",
                "description": "Delete user ECG value",
                "parameters": [
                    {
                        "name": "deviceId",
                        "in": "path",
                        "description": "The username of the person who want's ecg",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/ECGResponse"
                        }
                    },
                    "203": {
                        "description": "Non Authoritative Information",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "204": {
                        "description": "No Content Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorize",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "408": {
                        "description": "Timeout",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ecg/updatesymptom/{ecgid}": {
            "x-swagger-router-controller": "ecgController",
            "put": {
                "tags": [
                    "ECG"
                ],
                "operationId": "updatesymptom",
                "description": "Delete user ECG value",
                "parameters": [
                    {
                        "name": "ecgid",
                        "in": "path",
                        "description": "The username of the person who want's ecg",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "symbody",
                        "in": "body",
                        "description": "The username of the person who want's ecg",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/symbody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/ECGResponse"
                        }
                    },
                    "203": {
                        "description": "Non Authoritative Information",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "204": {
                        "description": "No Content Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorize",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "408": {
                        "description": "Timeout",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/get/stress/{stressid}": {
            "x-swagger-router-controller": "stressController",
            "get": {
                "tags": [
                    "Stress"
                ],
                "operationId": "getStressBystressid",
                "description": "Delete user ECG value",
                "parameters": [
                    {
                        "name": "stressid",
                        "in": "path",
                        "description": "The username of the person who want's ecg",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/StressResponse"
                        }
                    },
                    "203": {
                        "description": "Non Authoritative Information",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "204": {
                        "description": "No Content Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorize",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "408": {
                        "description": "Timeout",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stress/updatesymptom/{stressid}": {
            "x-swagger-router-controller": "stressController",
            "put": {
                "tags": [
                    "Stress"
                ],
                "operationId": "updatesymptomstress",
                "description": "Delete user ECG value",
                "parameters": [
                    {
                        "name": "stressid",
                        "in": "path",
                        "description": "The username of the person who want's ecg",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "symbody",
                        "in": "body",
                        "description": "The username of the person who want's ecg",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/symbody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/ECGResponse"
                        }
                    },
                    "203": {
                        "description": "Non Authoritative Information",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "204": {
                        "description": "No Content Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorize",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "408": {
                        "description": "Timeout",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "StressResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "required": [
                            "type",
                            "id"
                        ],
                        "properties": {
                            "type": {
                                "type": "string"
                            },
                            "id": {
                                "type": "string"
                            },
                            "attributes": {
                                "type": "object",
                                "properties": {
                                    "hr": {
                                        "type": "number"
                                    },
                                    "mrr": {
                                        "type": "number"
                                    },
                                    "sdnn": {
                                        "type": "number"
                                    },
                                    "symptoms": {
                                        "type": "string"
                                    },
                                    "rmssd": {
                                        "type": "number"
                                    },
                                    "pr": {
                                        "type": "number"
                                    },
                                    "qt": {
                                        "type": "number"
                                    },
                                    "minHr": {
                                        "type": "number"
                                    },
                                    "maxHr": {
                                        "type": "number"
                                    },
                                    "qtc": {
                                        "type": "number"
                                    },
                                    "qrs": {
                                        "type": "number"
                                    },
                                    "pnn50": {
                                        "type": "number"
                                    },
                                    "nn150": {
                                        "type": "number"
                                    },
                                    "lat": {
                                        "type": "string"
                                    },
                                    "long": {
                                        "type": "string"
                                    },
                                    "vmRatio": {
                                        "type": "string"
                                    },
                                    "nn2k": {
                                        "type": "number"
                                    },
                                    "deviceType": {
                                        "type": "string"
                                    },
                                    "fullReport": {
                                        "type": "boolean"
                                    },
                                    "value": {
                                        "type": "array",
                                        "items": {
                                            "type": "number"
                                        }
                                    },
                                    "rrInterval": {
                                        "type": "array",
                                        "items": {
                                            "type": "number"
                                        }
                                    },
                                    "arrHr": {
                                        "type": "array",
                                        "items": {
                                            "type": "number"
                                        }
                                    },
                                    "createdTs": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "firmwareVersion": {
                                        "type": "string"
                                    },
                                    "appVersion": {
                                        "type": "string"
                                    },
                                    "phoneModel": {
                                        "type": "string"
                                    },
                                    "osVersion": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "meta": {
                    "type": "object"
                },
                "include": {
                    "type": "object"
                }
            }
        },
        "symbody": {
            "type": "object",
            "properties": {
                "symptoms": {
                    "type": "string"
                }
            }
        },
...
        "ErrorResponse": {
            "title": "error",
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "title": {
                                "type": "string"
                            },
                            "code": {
                                "type": "string"
                            },
                            "detail": {
                                "type": "string"
                            },
                            "source": {
                                "type": "object",
                                "properties": {
                                    "pointer": {
                                        "type": "string"
                                    },
                                    "parameter": {
                                        "type": "string"
                                    }
                                }
                            },
                            "meta": {
                                "type": "object"
                            }
                        },
                        "required": [
                            "title",
                            "code",
                            "detail"
                        ]
                    }
                }
            }
        }
    }
}